payments/requirements.txt
certifi==2021.5.30
chardet==4.0.0
Click==8.0.1
Flask==2.0.3
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
stripe==6.0.0
toml==0.10.2
Werkzeug==2.0.1
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
python-jose[cryptography]>=3.3.0